Fixing part 2

+++ database/migrations/2021_11_11_153247_create_details_sales_table.php
@@ -16,7 +16,7 @@ class CreateDetailsSalesTable extends Migration
         Schema::create('details_sales', function (Blueprint $table) {
             $table->id();
             $table->double('cost');
-            $table->integer('quantiy');
+            $table->BigInteger('quantiy');
             $table->unsignedBigInteger('id_sales')->index('fk_id_sales');
             $table->unsignedBigInteger('id_product')->index('fk_id_product');
             $table->softDeletes();
